QGIM: documentation update
Document that we support reference wrappers and smart pointers for the
model. Turn the construction of the model into a dedicated section,
and add relevant snippets. Rename and reorganize sections a bit to
introduce basic concepts (rows, columns, multi-role items) first,
before the rather advanced topic of trees and value/pointer trade-offs
for rows.

Change-Id: Ia622783f9cd7aa345217186b438e5e883b9c25b0
Reviewed-by: Mate Barany <mate.barany@qt.io>