Add OffscreenSurface capability to QPlatfromIntegration
The capability reflects whether the function
QPlatfromIntegration::createPlatfromOffscreenSurface is available.

See QOffscreenSurface::create for details how offscreen surface
is handled. If no offscreen surface is created, it falls back
to an invisible window, using of which in non-ui thread causes
failures.

In QtMultimedia, we want to know in advance whether offscreen surface
can be created in non-ui thread, so we add the capability in the patch.

Pick-to: 6.8 6.10
Task-number: QTBUG-131107
Change-Id: I15e6b1b66aaac18020abca661415aea5c8779192
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>