QGIM: fix test coverage for gadgets-that-are-tuples
The test data for SingleColumn and MultiColumn disambiguation types
was incorrectly removed in 7803e6c000cc0fddba392bcbaefdc3a93ff4b26c.

Inserting rows of a MultiColumn wrapper around a pointer doesn't work,
so QEXPECT_FAIL that test for now.

As a drive-by, remove an obsolete QEXPECT_FAIL; we never add a table of
pointers in multiple value categories, as they all reference the same
data.

Change-Id: Ief4750094f8262f01ff81aa41d113971abd37307
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>