Remove QWindow argument from QWindowSystemInterface::handleThemeChange
The QPlatformTheme, and consumers on top such as QStyleHints, are app
global, so it doesn't make sense to notify that a single window has
changed its theme.

None of the QPA plugins pass a QWindow to the function today, after
the Windows plugin was taught a few new tricks in a preceding change.

[ChangeLog][QPA] QWindowSystemInterface::handleThemeChange no longer
takes an optional QWindow.

Change-Id: Ia761c95559a56a00763bd3cb95332f377e8de483
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>