Disable {position,resize}Automatic on QPlatformWindow creation
Once the platform window has been created it has had its chance to
set a default position and size, which is reflected back to the
QWindow. Destroying and recreating the window after that should
keep the position and size, even when those have been changed by
the end user via the system's window manager.

We were effectively doing this for size already, as initialGeomery
only applied a default size if the size was 0x0, but we were failing
to preserve the position. We now handle both cases explicitly.

[ChangeLog][QtGui] A destroyed and recreated QWindow will now maintain
its position and size, instead of allowing the platform window another
chance at setting a default position and size. Users of QWindow and
its consumers that reuses a single window for many "logical" windows
need to explicitly position and size the window for each "use".

Change-Id: Ib9a395295e6dfdc6db908e2c96be60046f462c30
Reviewed-by: Liang Qi <liang.qi@qt.io>