Handle Windows theme changes in central theme listener
Instead of having each window listen for theme changes we
now have a single theme listener window, similar to the
way we handle screen management.

This reduces the amount of redundant callbacks to QWSI
for theme changes somewhat, but Windows still emits
several redundant events even for a single window.
We want the theme change event to be synchronous, so
there is no obvious way to debounce these events,
besides the clearing of the message queue that we
already do in this patch.

Since we don't know exactly what part of the theme
changed we can't bail out of the theme change event
to QWSI based on the color scheme being the same, as
the accent color or other parts of the theme might
have changed.

Change-Id: I827fa50effadf8a8e674a03ddc72958c60310f48
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Zhao Yuhang <2546789017@qq.com>