QUrl: fix mismatch of encoded/decoded spaces in setXxx
Amends commit d064c26d2c1fb1f78013031a83b661ef6f74ce21, which introduced
qt_encodeFromUser(). As far as I can tell and test, only the space
character was an exception, due to this code in qt_urlRecode():

    if (!(encoding & QUrl::EncodeSpaces))
        actionTable[0] = DecodeCharacter; // decode

That meant that encoded spaces was an opt-in, not an opt-out, but is not
the default in the defaultActionTable. This was missed in the new
function.

Fixes: QTBUG-135949
Change-Id: Ic8d2adac5c32f858748bfffd5654aac3c7a2f782
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>