QUrl: change the default action for spaces to decoded
That matches the QUrl::ComponentFormattingOptions flags and avoids
modifying the memcpy'ed action table for the default. This applies to
everything we set, since we keep decoded spaces in the strings in
QUrlPrivate.

Amends commit 1b7e9dba75f18342911bc6954be3e754322f091f ("Change the
component formatting enum values so the default is zero") which changed
from DecodeSpaces to EncodeSpaces. And amends the previous commit.

Task-number: QTBUG-135949
Change-Id: I2a9e4ffc62ca4a75f675fffd366204c0a4ee7337
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>