macOS: Port from deprecated NSAppearance.currentAppearance
When the system appearance changes, it will be reflected through
NSApp.effectiveAppearance, but NSAppearance.currentAppearance
and NSAppearance.currentDrawingAppearance will remain as is for
some reason (perhaps as a compatibility for apps that were not
ready to handle dark mode).

The original way to deal with this was to explicitly set the
NSAppearance.currentAppearance to the new effective appearance,
which would take effect for the thread from that point on. But
this API has been deprecated, most likely because overriding it
globally could mess up views or logic that was not prepared for
dark mode).

The replacement API, NSAppearance performAsCurrentDrawingAppearance
is given a block, and will only override NSAppearance.currentAppearance
during the block, which isolates it from other fragile components.

We generally use NSApp.effectiveAppearance in our own code when
resolving dark mode, so we only need to wrap AppKit APIs we use
with performAsCurrentDrawingAppearance, including NSColor for our
palette, and NSView/NSControl drawing in our styles.

For the latter case we introduce a helper that we can also use
in the Qt Quick Mac style. This helper can also be the basis
of shared style functionality between Widgets and Quick going
forward.

Fixes: QTBUG-135789
Done-with: Volker Hilsheimer <volker.hilsheimer@qt.io>
Change-Id: I4d006b8b1fbbd7ff888da8b903066c183adad188
Reviewed-by: Doris Verria <doris.verria@qt.io>