Drop workaround to turn popups into main windows
On Wayland there are additional requirements when making popups to Qt.

In regular cases we can guess what the popup parent should be based on
which window last received input, but as it's a heuristic with obscure
enough timing events this can sometimes fail.

In this situation the current code creates a toplevel instead of a
regular popup, having a popup take focus with a window decoration is
rarely a good UX. Rather than making an incorrect window, the new
approach is to not create a popup over the wire and synthesise an
immediate close event. Application code should be able to handle this as
it's the same path if the popup was denied server side.

Pick-to: 6.9
Task-number: QTBUG-136343
Change-Id: Ia76c36a047e93f2469a8bf0f4b94642f05179124
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>