rhi: d3d12: Fix up sampling in the mipmap generator
Unclear if the results are different after this. The RenderDoc-based
investigations show the generated mip levels identical at first glance,
but the original logic is doing useless things: trying to rely on
mipmapping is futile when generating the mipmaps, while surely all
mip levels should be exposed to the shader that selects one based
on the level passed in from the C++ side.

(of course, exposing that single level via a view might just make
the level argument ignored or clamped, with the result of
SampleLevel() probably being the same regardless)

Change-Id: Ieae52fd45d19113e2d0bdab1736fb5d78cd55c3f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>