Add Qt styling, ids, and classes to string overview graphics
While many graphics in core use a light blue color, it does not fit to
Qt and to the general styling of the documentation. This patch changes
the color to the Qt green.

In the long run, I would like to have styling switching between light
and dark mode. This patch applies ids and styling so that it could be
switched from outside, if the html document this is embedded into
sets the correct tags. This will probably require some features in qdoc.

Task-number: QTBUG-133882
Change-Id: Iadce79a5f5a99014abe97c19d9e76efac3d49dad
Reviewed-by: Mate Barany <mate.barany@qt.io>