wayland: Make initial commit for surface state opt-in
After setting up the shell surface we setup the initial state and have
to commit the surface to apply the surface role in order to get a
configure event from the compositor.

This is not a requirement for subsurfaces, and can be skipped.

It is also problematic for some 3rd party shells, such as
ext-session-lock-v1, which has a different setup behaviour.

Task-number: QTBUG-136337
Change-Id: I9e7874ec2f4150e24eb4790384eacc4123f9552b
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>