Tweak a regex to only match what it's meant to
It's testing for either a .qm extension or the end of the file-name;
but it shouldn't match a .qml extension, so it should test the .qm is
at the end.  Thus it's a test for .qm-or-nothing at the end.
Amends commit 1f2ab46745af12e4776695c1b1a2523ac170368e

Change-Id: I3e47691bb7f576d21d61ce4988fc476421615add
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>