iOS: Update RLIMIT_STACK when running with user-main stack
When running the user's main() we retarget the stack pointer to a custom
stack, with half the size of the main thread stack (i.e 512K instead of
1MB). During execution of the user's main() we want code that reads
RLIMIT_STACK to be aware of the limited stack space, as this might impact
where memory guards are placed, or how stack overflow is detected, such
as in the V4 QML engine.

Once we return from the user's main() we restore back to the original
stack size limit. We also handle re-entry into app.exec() as well as
app termination, by always making sure RLIMIT_STACK reflects the available
stack space.

Unfortunately we can't do the same for pthread_get_stacksize_np, as once
the main thread has been created we can't tweak the stack size, and the
pthread_get_stacksize_np function does not query the rlimit on iOS, as
it does in some cases on macOS.

Change-Id: I48e92e762e125cff0f7da2003c7a1e796835d391
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>